requirements.txt
numpy
pandas
streamlit
openai
anthropic
python-dotenv